Fix stack smashing bug in Programs/gnat.c
XXX - this change needs to be reviewed.

The code used a macro, and subtracted a value like this:

    deg = (RadTo15Deg(vars->vel_angle) - 3) % 24;
    for (i = 0; i < 7; i++) {
      opt_dir[deg]--;
      [...]

This is OK, until the macro returns zero, and then subtracts three...

Suddenly the array index is negative, and some bits on the stack
get corrupted.

This pattern was used widely in the code base.  I changed it so the
code does the subtraction of the value before the values are
converted from radians to degrees.  I'm not sure that this the correct
thing to do, but it does fix the stack corruption problem.

    deg = (RadTo15Deg(vars->vel_angle - 3)) % 24;

Someone who understands the intent of the code better can decide
if the code should do what it does now, or do the prior math, and
fixup negative values to be zero.  Or fix up the values to be
something else.

Update the Programs/README.txt file to show this robot works OK now.