Support running on Mac OS X
In addition to the changes needed to get the code to compile
using 'bmake', the THREAD_SWAPCONTEXT implementation has been
worked over to operate properly on Mac OS X.

Rework THREAD_SWAPCONTEXT code to allocate stack space and hold a
pointer to that space before creating the context for each program
thread.  The OS X swapcontext() implementation changes the stack
pointer set inside the context, so we cannot free() the memory using
that pointer.  Change the teardown loop to free the stack buffer
as well as the thread buffer, so we shouldn't be leaking any memory
between program runs.

The THREAD_SWAPCONTEXT code now runs without crashing on both
FreeBSD and Mac OS X.