Remove recursive function call in Src/Programs/tagman.c
The robot would make a recursive call, that ultimately caused
the thread stackspace to be exhausted, which caused the game
to die.  In some game setup scenarios, this happened almost
immediately on startup.  Add a comment about this to the code
and disable the recursive call.

Also ANSIify a couple of tagman's functions that I missed previously.